requirements.txt
Flask==3.0.3
psycopg[binary]==3.2.13
python-dotenv==1.0.1
pandas==2.3.2
openpyxl==3.1.5
numpy==2.3.3
reportlab==4.2.2
